+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
     - Simple link support for string cell rendering (#1001)
     - ``generic_raw`` assay plugin (#1128)
     - Overriding assay plugin via assay comment (#1128)
+    - App alert for UI sheet cache update (#1000)
 
 Changed
 -------
@@ -35,6 +36,8 @@ Changed
 - **Irodsinfo**
     - Retrieve iRODS config from ``IRODS_ENV_CLIENT`` setting (#1221)
     - Display ``IRODS_ENV_CLIENT`` in siteinfo via ``info_settings``
+- **Samplesheets**
+    - Move ``TestSheetSyncBase`` into ``test_views_taskflow``
 
 Fixed
 -----